backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
PyPDF2==3.0.1
pdf2image==1.16.3
pdfplumber==0.10.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
python-dotenv==1.0.0
numpy
pandas
scikit-learn
joblib
